gcf/update-notification/requirements.txt
google-cloud-firestore==2.9.1
functions-framework==3.0.0
Flask>=1.0,<3.0
pyjwt==2.8.0
